* Note: Add method to get TPitch from note
